consultation_was_sheduled

+++ db/migrate/20230523071740_create_consultations.rb
@@ -2,7 +2,7 @@ class CreateConsultations < ActiveRecord::Migration[6.0]
   def change
     create_table :consultations do |t|
       t.references :user, null: false, foreign_key: true
-      t.integer  :patient_id
+      t.bigint  :patient_id
       t.boolean :open, default: true
       t.timestamps
     end
